docker/data_aggregation/requirements.txt
paho-mqtt
#pysqlite3
mariadb
manuf
pandas
